refactor(observability-webhooks): remove ObservabilityWebhook components
- Delete `EditObservabilityWebhook`, `ObservabilityWebhooks`, and related columns components.
- Migrate `DeleteObservabilityWebhookModal` to a shared location for reuse.
- Update routes to navigate webhooks settings path.
- Replace styled components with `SettingsFormCard` for UI consistency.
- Enhance `WebhooksColumns` with tooltip and copy button functionalities.
