+++ assets/src/components/settings/global/observability/Observability.tsx
@@ -1,22 +1,14 @@
 import styled from 'styled-components'
 
 import { SettingsPageHeader } from 'components/settings/Settings'
-import { SubTabs } from 'components/utils/SubTabs'
 import { Outlet, useOutletContext } from 'react-router-dom'
-
-const directory = [
-  { path: 'providers', label: 'Providers' },
-  { path: 'webhooks', label: 'Webhooks' },
-]
 
 export default function Observability() {
   const curContext = useOutletContext()
 
   return (
     <ObservabilityWrapperSC>
-      <SettingsPageHeader heading="Observability">
-        <SubTabs directory={directory} />
-      </SettingsPageHeader>
+      <SettingsPageHeader heading="Observability" />
       <Outlet context={curContext} />
     </ObservabilityWrapperSC>
   )
